backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
# API & CORS
djangorestframework>=3.15
django-cors-headers>=4.4
# Env & DB URL parsing
python-decouple>=3.8
dj-database-url>=2.3
python-dotenv>=1.0
# JWT verification
PyJWT[crypto]>=2.9
requests>=2.32
